+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Add missing private function `Test-DscObjectHasProperty`.
+
 ## [0.7.0] - 2020-05-02
 
 ### Added